backend/requirements.txt
flask==3.0.3
flask-jwt-extended==4.6.0
flask-cors==4.0.1
flask-limiter==3.7.0
flask-migrate==4.0.7
flask-sqlalchemy==3.1.1
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
redis==5.0.8
celery==5.4.0
gunicorn==22.0.0
gevent==24.2.1
cryptography==43.0.0
requests==2.32.3
pillow==10.4.0
numpy==2.0.1
python-dotenv==1.0.1
alembic==1.13.2
werkzeug==3.0.3
click==8.1.7
